refactor(test): extract domain test builders into shared Yumney.TestBuilders project
Builders previously lived in each *.Domain.Tests project under Builders/, which
meant Application/Infrastructure/Integration test projects couldn't reuse them
without an awkward cross-test-project ProjectReference. Pulling them into a
single non-test class library makes them available to every consumer.

Project layout:

    tests/Yumney.TestBuilders/
    ├── Recipes/   (RecipeBuilder, IngredientBuilder, StepBuilder,
    │              QuantityBuilder, TimingInfoBuilder)
    ├── Shopping/  (ShoppingListBuilder, ShoppingListItemBuilder)
    ├── MealPlan/  (WeeklyPlanBuilder)
    └── Users/     (AppUserProfileBuilder)

Namespaces become SmartSolutionsLab.Yumney.TestBuilders.{Recipes,Shopping,
MealPlan,Users}; the four *.Domain.Tests projects now reference TestBuilders
and import the new namespace in their test files.

The new project references all four *.Domain projects (it has to — each
builder's API exposes that module's domain types). It is *not* a test project
(no IsTestProject, no xUnit/Test.Sdk), just a thin library.

Verified: 274 + 182 + 131 + 243 = 830 Domain tests still pass.
