+++ Yumney.slnx
@@ -89,5 +89,6 @@
     <Project Path="tests/Yumney.Architecture.Tests/Yumney.Architecture.Tests.csproj" />
     <Project Path="tests/Yumney.Integration.Tests/Yumney.Integration.Tests.csproj" />
     <Project Path="tests/Yumney.Mcp.Server.Tests/Yumney.Mcp.Server.Tests.csproj" />
+    <Project Path="tests/Yumney.TestBuilders/Yumney.TestBuilders.csproj" />
   </Folder>
 </Solution>
