+++ tests/Yumney.Users.Domain.Tests/Yumney.Users.Domain.Tests.csproj
@@ -7,6 +7,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\src\Yumney.Users.Domain\Yumney.Users.Domain.csproj" />
+    <ProjectReference Include="..\Yumney.TestBuilders\Yumney.TestBuilders.csproj" />
   </ItemGroup>
 
   <ItemGroup>
